Redefine the PrimaryItem component as an independent component to avoid circular references
